testing rm line:37 uint8

+++ benchmark.py
@@ -34,7 +34,7 @@ class Benchmark:
         """Given a list of file names, return a list of images"""
         out = []
         for image in images:
-            out.append(cv2.imread(image).astype(np.uint8))
+            out.append(cv2.imread(image))
         return out
 
     def deprocess(self, image):
